+++ examples/CMakeLists.txt
@@ -1,3 +1,5 @@
+include_directories("../src/Physics")
+
 add_executable(rocket Rocket.cpp)
 
-target_link_libraries(rocket PUBLIC object math)
+target_link_libraries(rocket PUBLIC Physics Maths)
