+++ src/Common/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(Common SHARED
+	Size.hpp
+	Size.cpp
+)
