--- src/common/CMakeLists.txt
@@ -1,4 +1,0 @@
-add_library(common SHARED
-	Size.hpp
-	Size.cpp
-)
