+++ tests/Maths/Integration.GTest.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "../../src/maths/Integration.hpp"
+#include "Maths/Integration.hpp"
 
 namespace PhysEn {
 namespace GTest {
